Redefine SyntaxSQLException's vendor codes
