+++ shardingsphere-kernel/shardingsphere-single-table/shardingsphere-single-table-core/src/main/java/org/apache/shardingsphere/singletable/exception/UnsupportedDropCascadeTableException.java
@@ -28,6 +28,6 @@ public final class UnsupportedDropCascadeTableException extends SyntaxSQLExcepti
     private static final long serialVersionUID = -5995575055241171117L;
     
     public UnsupportedDropCascadeTableException() {
-        super(XOpenSQLState.FEATURE_NOT_SUPPORTED, 4, "DROP TABLE ... CASCADE is not supported");
+        super(XOpenSQLState.FEATURE_NOT_SUPPORTED, 3, "DROP TABLE ... CASCADE is not supported");
     }
 }
